KML parser working
KML parser will now add to swaths from KML and then add swaths to a planned flight. The "plannedFlight" will then be matched to a "project" where it will be matched with actual flight swaths.

+++ FlightPlanMatcher/FlightPlanMatcher/PlannedSwath.cs
@@ -6,10 +6,10 @@ namespace FlightPlanMatcher
 {
     class PlannedSwath
     {
-        public float StartLat { get; set; }
-        public float StartLong { get; set; }
-        public float EndLat { get; set; }
-        public float EndLong { get; set; }
+        public double StartLat { get; set; }
+        public double StartLong { get; set; }
+        public double EndLat { get; set; }
+        public double EndLong { get; set; }
 
         public int PlannedAltitude { get; set; }
         // flight run number according to KML flight plan
